Parse: Split for_expr rules into several
PLY doesn't allow deducing which grammar rule matched
and this often causes bugs.

We had a bug with `for` where all `for`s were `downto`
`for`s due to a string comparison which is not the responsibility
of the parser, but the lexer.

This fix makes it harder to introduce such bugs.