AST test: Regression test for __eq__ use of .type
__eq__ in AST Nodes assumes two objects of the same class
have the same attributes (no attributes are added in runtime).

This is not true with the current type table, because it modifies
ast.Constructor objects to add the `type` attribute. This regression
test ensures equality works between ast.Constructor objects that
are constructed manually and those that are passed through the type
table.

This is fixed in #25 and the regression test should pass.

This is related to the fact that semantic checks are currently done
before parsing is completed - this will change in future versions.