Main: Separate loggers for lexer and parser.
The splitting is not complete, since the typeTable reuses the parser's
logger; I see no point in fixing it, since the typeTable will
eventually be decoupled from the parser.

The joint logger is also retained the case where Parser.__init__ creates
a lexer on the fly, so that the user of the parser may be notified of
lexing errors.